regroup ressource hint tests
