--- test/build_preload_module/main.js
@@ -1,1 +1,0 @@
-await Promise.resolve(42)
