--- test/dev_server/dev_server_link_modulepreload/dev_server_link_modulepreload.html
@@ -1,18 +1,0 @@
-<html>
-  <head>
-    <title>Title</title>
-    <meta charset="utf-8" />
-    <link rel="icon" href="data:," />
-    <link rel="modulepreload" href="./dev_server_link_modulepreload.js" />
-  </head>
-
-  <body>
-    <script type="module" src="./dev_server_link_modulepreload.js"></script>
-    <script type="module">
-      // Let browser time to log the warning about preload link not used
-      await new Promise((resolve) => {
-        setTimeout(resolve, 5000)
-      })
-    </script>
-  </body>
-</html>
