+++ test/ressource_hints/modulepreload/main.js
@@ -1,5 +1,7 @@
 import { value } from "./file.js"
 
+await Promise.resolve(42)
+
 console.log(value)
 
 export { value }
