+++ test/ressource_hints/modulepreload/modulepreload_build.test.js
@@ -1,16 +1,11 @@
 import { assert } from "@jsenv/assert"
-import {
-  resolveUrl,
-  urlToRelativeUrl,
-  urlToBasename,
-  readFile,
-} from "@jsenv/filesystem"
+import { resolveUrl, urlToRelativeUrl, readFile } from "@jsenv/filesystem"
 
 import { buildProject } from "@jsenv/core"
 import { jsenvCoreDirectoryUrl } from "@jsenv/core/src/internal/jsenvCoreDirectoryUrl.js"
 import { GENERATE_ESMODULE_BUILD_TEST_PARAMS } from "@jsenv/core/test/TEST_PARAMS_BUILD_ESMODULE.js"
 import {
-  findNode,
+  findNodeByTagName,
   getHtmlNodeAttributeByName,
 } from "@jsenv/core/src/internal/compiling/compileHtml.js"
 
@@ -19,32 +14,32 @@ const testDirectoryRelativeUrl = urlToRelativeUrl(
   testDirectoryUrl,
   jsenvCoreDirectoryUrl,
 )
-const testDirectoryname = urlToBasename(testDirectoryRelativeUrl)
 const jsenvDirectoryRelativeUrl = `${testDirectoryRelativeUrl}.jsenv/`
 const buildDirectoryRelativeUrl = `${testDirectoryRelativeUrl}dist/esmodule/`
+const entryPointMap = {
+  [`./${testDirectoryRelativeUrl}modulepreload.html`]: "./main.html",
+}
 
 const { buildMappings } = await buildProject({
   ...GENERATE_ESMODULE_BUILD_TEST_PARAMS,
   jsenvDirectoryRelativeUrl,
   buildDirectoryRelativeUrl,
-  entryPointMap: {
-    [`./${testDirectoryRelativeUrl}${testDirectoryname}.html`]: "./main.html",
-  },
+  entryPointMap,
   // logLevel: "debug",
 })
 
-// ensure link.href is correct
+// ensure link.href is updated
 {
   const buildDirectoryUrl = resolveUrl(
     buildDirectoryRelativeUrl,
     jsenvCoreDirectoryUrl,
   )
   const htmlBuildUrl = resolveUrl("main.html", buildDirectoryUrl)
-  const htmlString = await readFile(htmlBuildUrl)
-  const linkPreload = findNode(htmlString, (node) => node.nodeName === "link")
-  const href = getHtmlNodeAttributeByName(linkPreload, "href").value
+  const htmlString = await readFile(htmlBuildUrl, { as: "string" })
+  const preloadLinkNode = findNodeByTagName(htmlString, "link")
+  const hrefAttribute = getHtmlNodeAttributeByName(preloadLinkNode, "href")
 
-  const actual = href
-  const expected = buildMappings[`${testDirectoryRelativeUrl}src/main.css`]
+  const actual = hrefAttribute.value
+  const expected = buildMappings[`${testDirectoryRelativeUrl}main.js`]
   assert({ actual, expected })
 }
