+++ test/ressource_hints/modulepreload_dependency/dependency.js
@@ -1,0 +1,1 @@
+export const value = 42
