+++ test/ressource_hints/modulepreload_dependency/main.js
@@ -1,0 +1,5 @@
+import { value } from "./dependency.js"
+
+console.log(value)
+
+export { value }
