+++ test/ressource_hints/modulepreload_dependency/modulepreload_dependency.html
@@ -5,7 +5,7 @@
     <meta charset="utf-8" />
     <!-- <link rel="icon" href="data:," /> -->
 
-    <link rel="modulepreload" href="./file.js" />
+    <link rel="modulepreload" href="./dependency.js" />
   </head>
 
   <body>
