+++ test/ressource_hints/preload_css/preload_css_build.test.js
@@ -5,7 +5,7 @@ import { buildProject } from "@jsenv/core"
 import { jsenvCoreDirectoryUrl } from "@jsenv/core/src/internal/jsenvCoreDirectoryUrl.js"
 import { GENERATE_ESMODULE_BUILD_TEST_PARAMS } from "@jsenv/core/test/TEST_PARAMS_BUILD_ESMODULE.js"
 import {
-  findNodeByTagName,
+  findNode,
   getHtmlNodeAttributeByName,
 } from "@jsenv/core/src/internal/compiling/compileHtml.js"
 
@@ -16,30 +16,29 @@ const testDirectoryRelativeUrl = urlToRelativeUrl(
 )
 const jsenvDirectoryRelativeUrl = `${testDirectoryRelativeUrl}.jsenv/`
 const buildDirectoryRelativeUrl = `${testDirectoryRelativeUrl}dist/esmodule/`
-const entryPointMap = {
-  [`./${testDirectoryRelativeUrl}build_modulepreload.html`]: "./main.html",
-}
 
 const { buildMappings } = await buildProject({
   ...GENERATE_ESMODULE_BUILD_TEST_PARAMS,
   jsenvDirectoryRelativeUrl,
   buildDirectoryRelativeUrl,
-  entryPointMap,
+  entryPointMap: {
+    [`./${testDirectoryRelativeUrl}preload_css.html`]: "./main.html",
+  },
   // logLevel: "debug",
 })
 
-// ensure link.href is updated
+// ensure link.href is correct
 {
   const buildDirectoryUrl = resolveUrl(
     buildDirectoryRelativeUrl,
     jsenvCoreDirectoryUrl,
   )
   const htmlBuildUrl = resolveUrl("main.html", buildDirectoryUrl)
-  const htmlString = await readFile(htmlBuildUrl, { as: "string" })
-  const preloadLinkNode = findNodeByTagName(htmlString, "link")
-  const hrefAttribute = getHtmlNodeAttributeByName(preloadLinkNode, "href")
+  const htmlString = await readFile(htmlBuildUrl)
+  const linkPreload = findNode(htmlString, (node) => node.nodeName === "link")
+  const href = getHtmlNodeAttributeByName(linkPreload, "href").value
 
-  const actual = hrefAttribute.value
-  const expected = buildMappings[`${testDirectoryRelativeUrl}main.js`]
+  const actual = href
+  const expected = buildMappings[`${testDirectoryRelativeUrl}src/main.css`]
   assert({ actual, expected })
 }
