+++ test/ressource_hints/preload_script/preload_script.html
@@ -3,11 +3,11 @@
     <title>Title</title>
     <meta charset="utf-8" />
     <link rel="icon" href="data:," />
-    <link rel="preload" href="./dev_server_preload_script.js" as="script" />
+    <link rel="preload" href="./script.js" as="script" />
   </head>
 
   <body>
-    <script type="module" src="./dev_server_preload_script.js"></script>
+    <script type="module" src="./script.js"></script>
     <script type="module">
       // Let browser time to log the warning about preload link not used
       await new Promise((resolve) => {
